break out weak deps tests
